move JSON nanobind casters into separate header file
